Cache referenced image files
Find any img tags, pull down a copy of their src and keep it locally,
and then rewrite the url to reference the local copy.

This has the benefit of truly snapshotting how my blog looks at a point
in time, at the expense of a much larger _deploy repo.

Idea for the future, where internet connections may be to shitty to
continue doing this; add a flag in the yaml front matter and do this
conditionally based on that so that already cached stuff continues to
use the cache, and new stuff can use live URLs until I find better
internet again.